Restructure component folders
